requirements.txt
streamlit==1.28.1
firebase-admin==6.4.0
beautifulsoup4==4.12.2
requests==2.32.2
groq==0.4.2
python-dotenv==1.0.1
pandas==2.2.3
plotly==5.17.0
lxml==4.9.3
